pyproject.toml
[tool.poetry]
name = "django_playground"
version = "0.1.0"
description = "Testing place for django apps"
authors = ["Mohamed Nesredin <m.n.kaizen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.2"
argon2-cffi = "^19.2"
confusable_homoglyphs = "^3.2"
celery = {extras = ["redis"], version = "^4.4"}
dj-database-url = "^0.5"
django = "^3.0"
djangorestframework = "^3.11"
django-axes = "^5.3"
django-cors-headers = "^3.2"
django-filter = "^2.2"
django-import-export = "^2.0"
django-storages = {extras = ["dropbox"], version = "^1.9"}
drf-yasg = "^1.17"
graphene-django = "^2.9"
gunicorn = "^20.0"
Pillow = "^7.0"
psycopg2-binary = "^2.8"
python-decouple = "^3.3"
pwned-passwords-django = "^1.4"
whitenoise = {extras = ["brotli"], version = "^5.0"}

[tool.poetry.dev-dependencies]
bandit = "^1.6"
black = "^19.10b0"
coverage = "^5.0"
django-coverage-plugin = "^1.8"
django-debug-toolbar = "^2.2"
factory-boy = "^2.12"
hypothesis = {extras = ["django"], version = "^5.6"}
isort = "^4.3"
mypy = "^0.770"
mkdocs = "^1.1"
mkdocs-material = "^4.6"
pre-commit = "^2.2"
pytest = "^5.4"
pytest-cov = "^2.8"
pytest-django = "^3.8"
pytest-sugar = "^0.9"
radon = "^4.1"
safety = "^1.8"
helium = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
